resilient rail working with RV_LA_WB

+++ x10.runtime/src-x10/x10/xrx/txstore/TxManagerForRail_RV_LA_WB.x10
@@ -21,7 +21,7 @@ public class TxManagerForRail_RV_LA_WB[K] {K haszero} extends TxManagerForRail[K
 
     public def this(data:TxRail[K], immediateRecovery:Boolean) {
         super(data, immediateRecovery);
-        if (here.id == 0) Console.OUT.println("TxManagerForRail_RL_EA_UL");
+        if (here.id == 0) Console.OUT.println("TxManagerForRail_RV_LA_WB");
     }
     
     public def getLocking(id:Long, index:Long):K {
@@ -33,6 +33,7 @@ public class TxManagerForRail_RV_LA_WB[K] {K haszero} extends TxManagerForRail[K
     }
     
     public def get(id:Long, index:Long):K {
+        if (TxConfig.TM_DEBUG) Console.OUT.println("Tx["+id+"] " + TxConfig.txIdToString (id)+ " here["+here+"] get rail["+index+"] ...");
         val log = txLogManager.getOrAddTxLog(id);
         try {
             log.lock(1);
@@ -40,13 +41,16 @@ public class TxManagerForRail_RV_LA_WB[K] {K haszero} extends TxManagerForRail[K
             val location = log.getLastUsedLocation();
             if (added)
                 data.logValueAndVersion(index, location, log);
-            return log.getCurrentValue(location);
-        }finally {
+            val v = log.getCurrentValue(location);
+            if (TxConfig.TM_DEBUG) Console.OUT.println("Tx["+id+"] " + TxConfig.txIdToString (id)+ " here["+here+"] get rail["+index+"] returns "+v+" ...");
+            return v;
+        } finally {
             log.unlock(1);
         }
     }
     
     public def put(id:Long, index:Long, newValue:K):void {
+        if (TxConfig.TM_DEBUG) Console.OUT.println("Tx["+id+"] " + TxConfig.txIdToString (id)+ " here["+here+"] put rail["+index+"]="+newValue+" ...");
         val log = txLogManager.getOrAddTxLog(id);
         try {
             log.lock(1);
@@ -56,15 +60,19 @@ public class TxManagerForRail_RV_LA_WB[K] {K haszero} extends TxManagerForRail[K
             log.getOrAddItem(index);
             val location = log.getLastUsedLocation();
             log.logPut(location, newValue);
-        }finally {
+            if (TxConfig.TM_DEBUG) Console.OUT.println("Tx["+id+"] " + TxConfig.txIdToString (id)+ " here["+here+"] put rail["+index+"]="+newValue+" done successfully...");
+        } finally {
             log.unlock(1);
         }
     }
     
     public def validate(id:Long):void {
+        if (TxConfig.TM_DEBUG) Console.OUT.println("Tx["+id+"] " + TxConfig.txIdToString (id)+ " here["+here+"] validate ...");
         val log = txLogManager.searchTxLog(id);
-        if (log == null || log.id() == -1)
+        if (log == null || log.id() == -1) {
+            if (TxConfig.TM_DEBUG) Console.OUT.println("Tx["+id+"] " + TxConfig.txIdToString (id)+ " here["+here+"] validate failed, log not found!!...");
             return;
+        }
         try {
             log.lock(1);
             val writeTx = log.validateRV_LA_WB(data);
@@ -73,45 +81,66 @@ public class TxManagerForRail_RV_LA_WB[K] {K haszero} extends TxManagerForRail[K
                     ensureActiveStatus();
                 log.writeValidated = true;
             }
+            if (TxConfig.TM_DEBUG) Console.OUT.println("Tx["+id+"] " + TxConfig.txIdToString (id)+ " here["+here+"] validate succeeded ...");
         } catch (ex:Exception) {
+            if (TxConfig.TM_DEBUG) {
+                Console.OUT.println("Tx["+id+"] " + TxConfig.txIdToString (id)+ " here["+here+"] validate failed  ex["+ex.getMessage()+"] will call abort...");
+                ex.printStackTrace();
+            }
             log.abortRV_LA_WB(data);
             txLogManager.deleteAbortedTxLog(log);
+            throw ex;
         } finally {
             log.unlock(1);
         }
     }
     
     public def commit(id:Long):void {
+        if (TxConfig.TM_DEBUG) Console.OUT.println("Tx["+id+"] " + TxConfig.txIdToString (id)+ " here["+here+"] commit ...");
         val log = txLogManager.searchTxLog(id);
-        if (log == null || log.id() == -1)
+        if (log == null || log.id() == -1) {
+            if (TxConfig.TM_DEBUG) Console.OUT.println("Tx["+id+"] " + TxConfig.txIdToString (id)+ " here["+here+"] commit failed, log not found!!!...");
             return;
+        }
         try {
             log.lock(1);
             log.commitRV_LA_WB(data);
+            if (TxConfig.TM_DEBUG) Console.OUT.println("Tx["+id+"] " + TxConfig.txIdToString (id)+ " here["+here+"] commit succeeded ...");
         } finally {
             log.unlock(1);
+            txLogManager.deleteTxLog(log);
         }
-        txLogManager.deleteTxLog(log);
     }
     
     public def abort(id:Long):void {
+        if (TxConfig.TM_DEBUG) Console.OUT.println("Tx["+id+"] " + TxConfig.txIdToString (id)+ " here["+here+"] abort ...");
         val log = txLogManager.searchTxLog(id);
-        if (log == null || log.id() == -1)
+        if (log == null || log.id() == -1) {
+            if (TxConfig.TM_DEBUG) Console.OUT.println("Tx["+id+"] " + TxConfig.txIdToString (id)+ " here["+here+"] abort failed, log not found ...");    
             return;
+        }
         try {
             log.lock(1);
             log.abortRV_LA_WB(data);
+            if (TxConfig.TM_DEBUG) Console.OUT.println("Tx["+id+"] " + TxConfig.txIdToString (id)+ " here["+here+"] abort succeeded ...");
         } finally {
             log.unlock(1);
+            txLogManager.deleteAbortedTxLog(log);
         }
-        txLogManager.deleteAbortedTxLog(log);
     }
 
     public def getTxCommitLog(id:Long):HashMap[Long,K] {
+        if (TxConfig.TM_DEBUG) Console.OUT.println("Tx["+id+"] " + TxConfig.txIdToString (id)+ " here["+here+"] getTxCommitLog ...");
         val log = txLogManager.searchTxLog(id);
-        if (log == null || log.id() == -1)
+        if (log == null || log.id() == -1) {
+            if (TxConfig.TM_DEBUG) Console.OUT.println("Tx["+id+"] " + TxConfig.txIdToString (id)+ " here["+here+"] getTxCommitLog failed, log not found !!!...");
             return null;
-        return log.getTxCommitLogRV_LA_WB();
+        }
+        val l = log.getTxCommitLogRV_LA_WB();
+        if (TxConfig.TM_DEBUG) {
+            Console.OUT.println("Tx["+id+"] " + TxConfig.txIdToString (id)+ " here["+here+"] getTxCommitLog succeeded, log = "+getCommitLogAsString(l)+" ...");
+        }
+        return l;
     }
     
     public def lockAll(id:Long, start:Long, opPerPlace:Long, indices:Rail[Long],readFlags:Rail[Boolean]):void {
@@ -120,5 +149,15 @@ public class TxManagerForRail_RV_LA_WB[K] {K haszero} extends TxManagerForRail[K
     
     public def unlockAll(id:Long, start:Long, opPerPlace:Long, indices:Rail[Long],readFlags:Rail[Boolean]):void {
         throw new Exception("operation not supported for baseline tx manager");
+    }
+    
+    private def getCommitLogAsString(log:HashMap[Long,K]) {
+        var str:String = "";
+        if (log != null && log.entries() != null) {
+            for (e in log.entries()) {
+                str += "("+e.getKey() + "," + e.getValue() + "):";
+            }
+        }
+        return str;
     }
 }
